// src/models/Product.ts
/* eslint-disable @typescript-eslint/naming-convention */
export interface Product {
    product_id?: number;
    product_name?: string;
    product_stock?: number;
    product_sellingprice?: number;
    product_adquisitionprice?: number;
    product_description?: string;
    product_category?: string;
};